Properly cast the system call input and outputs
On clang, this statement caused compiler errors instead of only
warnings.

Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #41

+++ __uk_init_tls.c
@@ -135,7 +135,7 @@ static int __uk_init_tp(void *p)
 	 * might happen.
 	 */
 	if (uk_thread_current()) {
-		td->tid = uk_syscall_r_set_tid_address(&td->tid);
+		td->tid = (int)uk_syscall_r_set_tid_address((long)&td->tid);
 		if (td->tid < 0)
 			td->tid = 0;
 	} else {
